+++ src/components/TimePicker.jsx
@@ -69,13 +69,12 @@ const defaultProps = {
 class TimePicker extends React.PureComponent {
   constructor(props) {
     super(props);
-    const {focused, meridiem, time, timeMode, timezone, onTimezoneChange} = props;
-    const timeData = timeHelper.time(time, meridiem, timeMode, timezone);
+    const {focused, timezone, onTimezoneChange} = props;
+    const timeData = this.timeData();
     const timezoneData = timeHelper.tzForName(timeData.timezone)
 
     this.state = {
       focused,
-      timeData,
       timezoneData
     };
 
@@ -90,6 +89,12 @@ class TimePicker extends React.PureComponent {
     if (!timezone) {
       onTimezoneChange(timezoneData);
     }
+  }
+
+  timeData() {
+    const {meridiem, time, timeMode, timezone} = this.props;
+    const timeData = timeHelper.time(time, meridiem, timeMode, timezone);
+    return timeData;
   }
 
   languageData() {
@@ -112,7 +117,7 @@ class TimePicker extends React.PureComponent {
 
   getHourAndMinute() {
     const {timeMode} = this.props;
-    const {timeData} = this.state;
+    const timeData = this.timeData();
     // Since someone might pass a time in 24h format, etc., we need to get it from
     // timeData to 'translate' it into the local format, including its accurate meridiem
     const hour = (parseInt(timeMode) === 12) ? timeData.hour12 : timeData.hour24;
@@ -129,7 +134,7 @@ class TimePicker extends React.PureComponent {
   handleHourChange(hour) {
     hour = timeHelper.validate(hour);
     const {onHourChange} = this.props;
-    const [_, minute] = this.getHourAndMinute();
+    const minute = this.getHourAndMinute()[1];
     onHourChange && onHourChange(hour);
     this.handleTimeChange(`${hour}:${minute}`);
   }
@@ -137,7 +142,7 @@ class TimePicker extends React.PureComponent {
   handleMinuteChange(minute) {
     minute = timeHelper.validate(minute);
     const {onMinuteChange} = this.props;
-    const [hour, _] = this.getHourAndMinute();
+    const hour = this.getHourAndMinute()[0];
     onMinuteChange && onMinuteChange(minute);
     this.handleTimeChange(`${hour}:${minute}`);
   }
@@ -162,7 +167,7 @@ class TimePicker extends React.PureComponent {
 
   get meridiem() {
     const {meridiem} = this.props;
-    const {timeData} = this.state;
+    const timeData = this.timeData();
     const localMessages = this.languageData();
     // eslint-disable-next-line no-unneeded-ternary
     const m = (meridiem) ? meridiem : timeData.meridiem;